app/requirements.txt
flask==3.0.3
flask-login
mysql-connector
flask-wtf